add dependencies to utils
